% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.FacileFseaTtestAnalysisResult.R
\name{compare.FacileFseaTtestAnalysisResult}
\alias{compare.FacileFseaTtestAnalysisResult}
\title{Compare two (or more) ffsea results.}
\usage{
compare.FacileFseaTtestAnalysisResult(
  x,
  y,
  ...,
  method = NULL,
  score = c("logFC", "t", "NES"),
  features = NULL
)
}
\arguments{
\item{...}{you can provide more FacileFseaTtestAnalysisResult objects}

\item{features}{The features used to calucate the effect size of the gsea
result. This can be either \code{"all"}, or \code{"leading_edge"} to have this method
pull the genes, or a long data.frame with collection,name,feature_id to
pull these out manually using some externally specified approach.}

\item{genesets}{The genesets to show. If \code{NULL} (default), takes the \code{ntop}
genesets by pvalue. Users can provide a collection,name\link{,group} tibble
that pulls out specific genesets, and optionally groups them by
user-defined categories.}
}
\description{
This only works with ffsesa results generated from a ttest dge analysis
}
