% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_shifts.R
\name{geneset_shifts}
\alias{geneset_shifts}
\title{Generate geneset shift plots to compare N contrasts}
\usage{
geneset_shifts(
  genesets,
  ...,
  gsea.method = NULL,
  colors = NULL,
  ymin = 0.005,
  ymax = 0.025,
  bw.bg = 0.7,
  bw.gs = 0.9,
  ribbon_wrap_n = 18,
  legend.position = "none",
  xlims = NULL,
  with_pvals = TRUE
)
}
\arguments{
\item{genesets}{character vector of geneset names to extract from the
GeneSetDb objects from each results.}

\item{...}{arbitrary number of named ffsea results}

\item{colors}{aaaaarrrghhhhhhhhh}

\item{stats}{the name of the gsea result to pull pvalues from}
}
\description{
This funciton is used primarily to generate geneshift plots to compare
the AppSAA model vs 5XFAD, but why not generalize ...
}
\details{
We assume that the ffsea results each have a GeneSetDb that has genesets
by the same name in them.

The row and column order are determined by the order in which ffsea results
and genesets are provided in the respective variables.
}
