% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_simplify.R
\name{go_simplify}
\alias{go_simplify}
\title{Simplifies ffsea GO analyses}
\usage{
go_simplify(
  x,
  method = NULL,
  ontology = c("BP", "MF", "CC"),
  direction = c("all", "up", "down"),
  similarity = c("Resnik", "Lin", "Rel", "Jiang", "Wang"),
  threshold = 0.7,
  max_padj = 0.1,
  min_logFC = NULL,
  score_by = NULL,
  semdata = NULL,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{x}{A FacileFseaAnalysisResult}

\item{ontology}{which ontology are were summarizing? BP, MF, or CC}

\item{direction}{focus on all genes, or only up or down regulated genes
(works for ttest, not ANOVA)}

\item{similarity}{The method to use to measure similarity,
\href{https://www.bioconductor.org/packages/release/bioc/vignettes/GOSemSim/inst/doc/GOSemSim.html#semantic-similarity-measurement-based-on-go}{see here}.}

\item{min_logFC, max_padj}{parameters used to define signficant genes. If
the TREAT framework was used, defaults to the threshold tested, otherwise
its zero.}
}
\description{
We use rrvgo, goSemSim and friends
}
