% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea-utils.R
\name{leading_edge_scores}
\alias{leading_edge_scores}
\title{Aggregates gene-level stats from genes in leading edge.}
\usage{
leading_edge_scores(
  x,
  leading_edge = NULL,
  aggregate = c("t", "logFC"),
  stats = NULL,
  ...
)
}
\arguments{
\item{x}{The FacileFseaAnalysisResult objects where \code{fgsea} was
ran.}

\item{leading_edge}{Precaculated leading edge genes, as in the output from
\code{\link[=leading_edge]{leading_edge()}}. If \code{NULL}, the leading_edge will be caluclated for all
genesets tested \code{x} then scored. The scoring can take a long time if there
are many, so you may want to provide a filtered leading_edge set.}

\item{aggregate}{the column names from the dge results to summarize, defaults
to \code{c("t", "logFC")}}

\item{stats}{A name of a gsea method run in \code{"x"} to append pvalues from.
If \code{NULL} (default), none will be added.}
}
\value{
a tibble of geneset scores, aggregated by leading edge genesets
for columns in \code{aggregate}. If \code{"fgsea"} is a method run in \code{"x"}, \code{ES},
and \code{NES} columns will be returned as well.
}
\description{
The logFC and t-statistics of the genes from the leading edge
(defined in \code{\link[=extract_leading_edge]{extract_leading_edge()}}) are used to give an effect size of the
geneset shift.
}
\details{
This is used in place of the ES/NES mojo.
}
