% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-analysis.R
\name{pin_read_auto_analysis}
\alias{pin_read_auto_analysis}
\title{Restore the pinned analysis}
\usage{
pin_read_auto_analysis(
  board,
  name,
  key = NULL,
  definitions = NULL,
  elements = NULL,
  fds. = NULL,
  ...
)
}
\value{
a list of analysis results, with \verb{$name} \verb{$dge}, \verb{$gsea}, \verb{$go}
elements.
}
\description{
Restore the pinned analysis
}
