% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-analysis.R
\name{pin_write_auto_analysis}
\alias{pin_write_auto_analysis}
\title{Save the analysis result to a pinboard}
\usage{
pin_write_auto_analysis(
  board,
  x,
  name = NULL,
  key = NULL,
  elements = NULL,
  ...
)
}
\arguments{
\item{board}{the pinboard}

\item{x}{the FacileAutoanalysis result}

\item{name}{the name of the objec to pin. Defaults to \code{x$name}.}

\item{key}{a unique string that prefixes all of the objects serialized by
this project}
}
\description{
Save the analysis result to a pinboard
}
