% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-analysis.R
\name{run_analyses}
\alias{run_analyses}
\title{Executes the predefined analysis defined in an experiment setup table}
\usage{
run_analyses(
  definitions,
  board = pins::board_temp(),
  key = NULL,
  gdb = NULL,
  ffsea_methods = NULL,
  semdata = NULL,
  BPPARAM = BiocParallel::bpparam(),
  ...
)
}
\arguments{
\item{definitions}{path to the comparison definition file}

\item{board}{the pinboard to serialize the results to}
}
\description{
Executes the predefined analysis defined in an experiment setup table
}
\examples{
efds <- FacileData::exampleFacileDataSet()
esamples <- FacileData::with_sample_covariates(FacileData::samples(efds))
adefs <- read.csv(
  system.file("extdata", "example-auto-analysis.csv",
              package = "FacileIncubator"))
results <- run_analyses(adefs, samples = esamples)
}
