% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-analysis.R
\name{run_comparison}
\alias{run_comparison}
\title{Performs dge, gsea, and GO simplification in one shot for a given analysis.}
\usage{
run_comparison(
  params,
  samples = NULL,
  board = pins::board_temp(),
  name = NULL,
  gdb = NULL,
  ffsea_methods = NULL,
  go_ontologies = c("BP", "MF", "CC"),
  go_simplify_method = "ora",
  go_simplify_sem_sim_method = "Resnik",
  go_simplify_max_padj = 0.1,
  go_simplify_sem_threshold = 0.6,
  go_semdata = NULL,
  BPPARAM = BiocParallel::bpparam(),
  direct_dge = NULL,
  ...
)
}
\arguments{
\item{params}{a list of params used for the analysis. This is typically a
single row of a comparison table}

\item{samples}{the facile_frame that includes the sample info}

\item{board}{the pinboard to serialize the results to. If you don't wan to
use a fancy remote one, you can always use a local directory.}
}
\description{
Performs dge, gsea, and GO simplification in one shot for a given analysis.
}
