% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fheatmap.R
\name{fheatmap}
\alias{fheatmap}
\title{Approximate a facile heatmap function}
\usage{
fheatmap(x, gdb = NULL, rename_rows = NULL, ...)
}
\description{
This just wraps the \code{\link[sparrow:mgheatmap2]{sparrow::mgheatmap2()}} function, but can accept a
\code{facile_frame}
}
\details{
TODO: support batch correction over data retrieved from \code{x}
}
