% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_match.R
\name{template_match}
\alias{template_match}
\title{Score genes against a template of expression matching.}
\usage{
template_match(
  x,
  template = "ascending",
  cor_method = c("spearman", "pearson", "kendall"),
  ...
)
}
\arguments{
\item{x}{A FacileAnalysis result (likely FacileAnovaAnalysisResult).
We assume the formula was an intercept/effect model.}

\item{template}{Defines the template to match against. Can be specified by
name for simple template (like \code{"ascending"} or \code{"descending"}), or a
numeric vector that provides the pattern, ie. c(1, 2, 3, 2, 1) for a
"peak". Default \code{"ascending"}.}
}
\value{
a tibble of features that match the template pattern
}
\description{
Provides scores for genes where increasing values indicate closer
concordance to some pattern of interest such as genes that have monotonically
increasing, decreasing, or a peak in the middle, etc. This is called
"template matching" in the field, which was (I think) originally coined by
Pavlidis.
}
\details{
This implementation follows the main idea from the \code{PatternHunger} function
in the MApckg package.

Other resources:
\itemize{
\item Analysis of strain and regional variation in gene expression in mouse brain
https://doi.org/10.1186/gb-2001-2-10-research0042 (pavlidis et al.)
\item Gene Time Eχpression Warper: a tool for alignment, template matching and
visualization of gene expression time series.
https://doi.org/10.1093/bioinformatics/bti787
\item Post hoc pattern matching: assigning significance to statistically defined
expression patterns in single channel microarray data.
https://doi.org/10.1186/1471-2105-8-240
}

This is a bad name for this function, but just putting it down here.

GOTO: FacileAnalysis
}
\section{Usage Examples}{

Imagine we have several timepoints encoded in a 'group' covariate.
you can run an nova to find genes that have some association w/ time,
then find ones within that result that match your template (let's say
ones that increase over time\if{html}{\out{<div class="r">}}\preformatted{increasing <- fds \%>\%
  flm_def("group") \%>\%
  fdge() \%>\%
  template_match("ascending)
}\if{html}{\out{</div>}}
}

\references{
\url{https://rdrr.io/github/flajole/MApckg/man/PatternHunter.html}
}
